backend/server/requirements.txt
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
dnspython==1.15.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SocketIO==2.9.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.0
pymongo==3.6.1
python-engineio==2.0.4
python-socketio==1.9.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
